requirements.txt
Appium-Python-Client==2.1.2
